# Conflicts:
#	RTA-Rubric.xlsx

+++ FAST_RTA_SOLUTION/FAST_RTA_SOLUTION/BindPose.cpp
@@ -16,12 +16,12 @@ const DirectX::XMFLOAT4X4 * BindPose::GetBindPose() const
 	return &bindposeInverses[0];
 }
 
-int BindPose::GetNumBones() const
+unsigned long BindPose::GetNumBones() const
 {
 	return bindposeInverses.size();
 }
 
-void BindPose::init(int _num, const DirectX::XMFLOAT4X4 * _mat)
+void BindPose::init(unsigned long _num, const DirectX::XMFLOAT4X4 * _mat)
 {
 	bindposeInverses.clear();
 	bindposeInverses.resize(_num);
